Updated liboil snapshot to 0.3.15. Added some additional build fixes on win32. Build tested on win32, linux32.
Merged revisions 9899 via svnmerge from 
https://svn.songbirdnest.com/vendor/upstream/liboil

........
  r9899 | preed | 2008-09-02 11:22:34 -0700 (Tue, 02 Sep 2008) | 3 lines
  
  Vendor drop of the liboil library, version 0.3.15, taken from
  http://liboil.freedesktop.org/download/liboil-0.3.15.tar.gz; md5sum
  11dd39b1ca13ce2e0618d4df8303f137
........


git-svn-id: http://publicsvn.songbirdnest.com/vendor/trunk@9900 37441250-4157-47a9-a2c9-f9fe3212d2d3

+++ liboil/liboil/ref/diffsquaresum_f32.c
@@ -59,8 +59,8 @@ diffsquaresum_f32_ref(float *dest, float *src1, int sstr1, float *src2,
 	int i;
 
 	for(i=0;i<n;i++){
-		x = OIL_GET(src1, i*sstr1, double) -
-                  OIL_GET(src2, i*sstr2, double);
+		x = OIL_GET(src1, i*sstr1, float) -
+                  OIL_GET(src2, i*sstr2, float);
 		x = x*x;
 		tmp = sum;
 		sum += x;
